requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pyserial==3.5
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
